LPS-147831 Refactor SelectFolderButton to remove metal dependency
