+++ modules/apps/journal/journal-web/src/main/resources/META-INF/resources/move_articles_and_folders.jsp
@@ -190,5 +190,5 @@ JournalMoveEntriesDisplayContext journalMovesEntriesDisplayContext = new Journal
 			"selectFolderURL", selectFolderURL
 		).build()
 	%>'
-	module="js/SelectFolderButton.es"
+	module="js/SelectFolderButton"
 />
